Create index.html
This is part of a wider project, my website, and explores relative and absolute positioning along with CSS Less for styling. The idea being to graphically present my work and education history with more details on a hover event and touches on CSS specificity within the context of the project.